backend/kashmir-tourism-prediction-main/requirements.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
pyyaml>=6.0
tqdm>=4.65.0
